+++ backend/core/models.py
@@ -40,6 +40,7 @@ class Device(models.Model):
     last_port_scan = models.DateTimeField(blank=True, null=True)
     missed_scans = models.PositiveIntegerField(default=0)
     known = models.BooleanField(default=False)
+    is_gateway = models.BooleanField(default=False)
 
     class Meta:
         ordering = ["-online", "name", "ip"]